requirements.txt
discord.py
PyNaCl
python-dotenv
youtube-dl
youtube-search-python
requests